increase canvas rendering performance by lowering the call count to putImageData and sleeping for a bit before we flushd isplay changes.
Prior to this diff we were using more cycles than we needed by calling putImageData every time we received a display change from the server. Now we only call it when we go to redraw the canvas.

This alone caused the game to speed up to the point where it was unplayable, so now we sleep on the server for 3 ms before pushing a display change to the client. 3 ms is kind of arbitrary, it just seemed right on my laptop. It might not work for over the Internet or mobile